Disable swap and backup file generation
This keeps all text in memory which is no problem for most files. Big
files need to be handled with care and file recovery will be impossible.
Yet we don't want annoying swap files laying around. For the same reason
disable back up files.

Do not make a backup before overwriting a file. Most of the files being
edited are tracked in Git anyhow, vim does not have to create these
backup files!