Display buffers and status
vim-airline is a lean & mean status/tabline for vim that's light as air.

The buffers are displayed with a number to easier switch between them.
The color is selected to match the solarized color scheme. Powerline
fonts are enabled to display fancy icons.

Since the mode is displayed in the airline, we don't need to display it
anymore in a message.