Configure for planck ez keyboard
Using backslash as a leader is not useful for my needs, as backslash is
set using a modifier and pressing the delete key, which makes me stretch
the right small finger that much, that the hand is slightly misplaced
for most of the key sequences. Using "," is much more convenient.

Also map the F-keys (which are located at the home position of the left
hand) to move fast between buffers and the "controlled" keys at the home
position of the right hand to move between splits. This requires the
.zshrc to set `ssty -ixon`.

Also allow to edit and source the vimrc file with more leader mappings.

Map the leader before defining the mappings to allow using the
configured leader.